package: ts7553v2-utils: Remove hash, clone master for build
The ts7553v2-utils packages was locked in to a specific commit.
This was the preferred way of doing this in buildroot, so that
buildroot builds can have consistent output every time. However,
this leads to a lot of tail chasing as small fixes are applied to
utilities over time. Now, build from master branch, always.